refactor(elixir): extract secure channel modules into its own files
rename Ockam.Identity.SecureChannel.IdentityChannelMessage
to Ockam.Identity.SecureChannel.HandshakeMessage

Closes #3650
