+++ implementations/elixir/ockam/ockam/lib/ockam/identity/secure_channel/handshake_message/request.ex
@@ -1,0 +1,6 @@
+defmodule Ockam.Identity.SecureChannel.HandshakeMessage.Request do
+  @moduledoc """
+  Identity channel handshake request
+  """
+  defstruct [:contact, :proof]
+end
