+++ implementations/elixir/ockam/ockam/lib/ockam/identity/secure_channel/handshake_message/response.ex
@@ -1,0 +1,6 @@
+defmodule Ockam.Identity.SecureChannel.HandshakeMessage.Response do
+  @moduledoc """
+  Identity channel handshake response
+  """
+  defstruct [:contact, :proof]
+end
